Add testcase for converting from old highscores DB
